change joincolumn name to integer

+++ packages/core/database/lib/metadata/relations.js
@@ -423,13 +423,13 @@ const createJoinTable = (metadata, { attributeName, attribute, meta }) => {
         type: 'increments',
       },
       [joinColumnName]: {
-        type: 'float',
+        type: 'integer',
         column: {
           unsigned: true,
         },
       },
       [inverseJoinColumnName]: {
-        type: 'float',
+        type: 'integer',
         column: {
           unsigned: true,
         },
